add series

+++ src/UseCases/Notifications/ReadIntervalsUpdatedNotification.cs
@@ -3,4 +3,4 @@ using Telegram.Bot.Types;
 
 namespace UseCases.Notifications;
 
-public record ReadIntervalsUpdatedNotification(DateOnly Date, ChatId ChatId, long MessageId) : INotification;
+public record ReadIntervalsUpdatedNotification(DateOnly Date, ChatId ChatId, int MessageId) : INotification;
